CreateResidents
